SFA: Move model loader to model.ts
